requirements.txt
asgiref==3.8.1
astral==3.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-compressor==4.4
django-crispy-forms==2.1
django-filter==23.1
django-jazzmin==2.6.0
django-leaflet==0.30.0
django-widget-tweaks==1.4.12
gunicorn==21.2.0
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
requests==2.31.0
rjsmin==1.2.1
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.6.0
